+++ tests/ConfigTest.php
@@ -56,7 +56,7 @@ class ConfigTest extends TestCase
     }
 
     /**
-     * @see Mail_Helper::getSMTPSettings does this weird settype:
+     * @see Mail_Helper::getSMTPSettings did this weird settype:
      * settype($config['smtp']['auth'], 'boolean');
      * that does not work (Indirect modification error),
      * so test version that works