+++ ChangeLog.md
@@ -6,6 +6,7 @@
 - Fix bug creating new priority (@balsdorf)
 - add MailTransport class to encapsulate smtp transport (@glensc, #236, #234)
 - Add option to control ordering of custom field options (@balsdorf, #240)
+- Check for release name uniqueness (@balsdorf, #246)
 
 [3.1.10]: https://github.com/eventum/eventum/compare/v3.1.9...master