+++ ChangeLog.md
@@ -7,6 +7,7 @@
 - add MailTransport class to encapsulate smtp transport (@glensc, #236, #234)
 - Add option to control ordering of custom field options (@balsdorf, #240)
 - Check for release name uniqueness (@balsdorf, #246)
+- Add a per project recipient flag for outgoing emails (@balsdorf, #247)
 
 [3.1.10]: https://github.com/eventum/eventum/compare/v3.1.9...master